Nova Sonic Speech (Python)/requirements.txt
fastapi
uvicorn
pyaudio
python-dotenv
aws-sdk-bedrock-runtime  # or the correct name if available from AWS SDK
